fix: Credit 카드 결제 시 자동으로 결제 내역 생성 및 조회 기능 수정

+++ src/main/java/server/domain/credit/domain/CreditHistory.java
@@ -19,9 +19,9 @@ public class CreditHistory {
 
     private LocalDateTime createdAt;
 
-    private Long amount;
+    private Integer amount;
 
-    private Long amountSum;
+    private Integer amountSum;
 
     private String name;
